refactor country and city selection logic into separate module
