+++ settings.json
@@ -1,10 +1,10 @@
 {
-  "city": "Tunis",
-  "country": "Tunisia",
+  "city": "Quebec",
+  "country": "Canada",
   "theme": "navy",
   "language": "en",
   "position": {
-    "x": 1216,
-    "y": 248
+    "x": 718,
+    "y": 132
   }
 }
